# Improve transform allocation overhead (with focus on AutoSize)