--- a/osu.Framework/Graphics/Transforms/Transformable.cs
+++ b/osu.Framework/Graphics/Transforms/Transformable.cs
@@ -42,7 +42,12 @@ namespace osu.Framework.Graphics.Transforms
         /// </summary>
         public IEnumerable<Transform> Transforms => targetGroupingTrackers.SelectMany(t => t.Transforms);
 
-        internal IEnumerable<Transform> TransformsForTargetMember(string targetMember) =>
+        /// <summary>
+        /// Retrieves the <see cref="Transform"/>s for a given target member.
+        /// </summary>
+        /// <param name="targetMember">The target member to find the <see cref="Transform"/>s for.</param>
+        /// <returns>An enumeration over the transforms for the target member.</returns>
+        public IEnumerable<Transform> TransformsForTargetMember(string targetMember) =>
             getTrackerFor(targetMember)?.Transforms ?? Enumerable.Empty<Transform>();
 
         /// <summary>