A collection of a few low-effort allocation optimisations. Most applicable to the edge case where `AutoSizeDuration > 0` and size is updated every frame. This happens at the osu! menu.

I think a more permanent solution would be using a Lerp call in place of this, but we will need to obsolete / remove `AutoSizeEasing` (not sure we want this - it can give correct results in the case autosize is only changed once in a while... but maybe it's just not worth supporting?).

before

|                       Method |       Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |-----------:|--------:|--------:|------:|------:|------:|----------:|
|                     RunFrame |   739.0 us | 5.82 us | 4.86 us |     - |     - |     - |       1 B |
|         RunFrameWithAutoSize | 1,321.7 us | 6.70 us | 6.27 us |     - |     - |     - |     124 B |
| RunFrameWithAutoSizeDuration | 1,102.9 us | 5.52 us | 4.61 us |     - |     - |     - |     828 B |


after (enumeration reduction https://github.com/ppy/osu-framework/commit/246502b3becbac03fb810b3dab533f4a8b842819)

|                       Method |       Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |-----------:|--------:|--------:|------:|------:|------:|----------:|
|                     RunFrame |   721.0 us | 4.81 us | 4.26 us |     - |     - |     - |      10 B |
|         RunFrameWithAutoSize | 1,319.9 us | 5.05 us | 4.22 us |     - |     - |     - |       1 B |
| RunFrameWithAutoSizeDuration | 1,107.1 us | 7.04 us | 6.59 us |     - |     - |     - |     524 B |

after (list -> queue https://github.com/ppy/osu-framework/commit/dcaabb1a42ac60f5deee89b81ef9ddcd37f2b179)

|                       Method |       Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |-----------:|---------:|---------:|------:|------:|------:|----------:|
|                     RunFrame |   713.9 us |  3.07 us |  2.87 us |     - |     - |     - |      10 B |
|         RunFrameWithAutoSize | 1,319.8 us |  6.13 us |  5.73 us |     - |     - |     - |       1 B |
| RunFrameWithAutoSizeDuration | 1,113.2 us | 19.86 us | 17.61 us |     - |     - |     - |     492 B |

after (start list at size 1 https://github.com/ppy/osu-framework/commit/9de51768387e37a48d895728c91dbd6954311ed5)

|                       Method |       Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------------- |-----------:|--------:|--------:|------:|------:|------:|----------:|
|                     RunFrame |   712.6 us | 2.01 us | 1.78 us |     - |     - |     - |       1 B |
|         RunFrameWithAutoSize | 1,309.5 us | 4.96 us | 4.64 us |     - |     - |     - |       1 B |
| RunFrameWithAutoSizeDuration | 1,103.3 us | 8.26 us | 6.89 us |     - |     - |     - |     449 B |